added modular js index file to register and login pages, made access token save on login, file cleanup
