--- index.mjs
@@ -1,1 +1,0 @@
-import * as registerUser from "./src/js/register/registerUser.mjs";
