+++ login/index.html
@@ -23,14 +23,14 @@
     <main>
       <h1 class="text-center text-bg-dark p-3">Welcome to SocialDock</h1>
       <h2 class="text-center text-bg-info text-white my-3 p-3">Log in</h2>
-      <form method="post" action="login" class="d-flex flex-column gap-3 col-8 col-md-4 col-lg-5 col-xl-5 col-xxl-5 m-auto">
+      <form method="post" id="login-form" class="d-flex flex-column gap-3 col-8 col-md-4 col-lg-5 col-xl-5 col-xxl-5 m-auto">
         <div class="form-group">
           <label for="post-title">Email</label>
-          <input type="email" class="form-control" id="email" placeholder="example@noroff.no" />
+          <input type="email" class="form-control" id="email" name="email" placeholder="example@noroff.no" />
         </div>
         <div class="form-group">
           <label for="password">Password</label>
-          <input type="password" class="form-control" id="password" />
+          <input type="password" class="form-control" id="password" name="password" />
         </div>
         <button type="submit" class="btn btn-primary">Log in</button>
         <p>Don't have an account? <a href="/register/">Register here...</a></p>
@@ -39,6 +39,6 @@
     <footer>
       <p class="text-white-50 text-center text-bg-dark m-0 mt-4 p-5">SocialDock &#169; 2022</p>
     </footer>
-    <script src="/src/index.mjs" type="module"></script>
+    <script src="../src/js/login/index.mjs" type="module"></script>
   </body>
 </html>
