+++ register/index.html
@@ -23,7 +23,7 @@
     <main>
       <h1 class="text-center text-bg-dark p-3">Welcome to SocialDock</h1>
       <h2 class="text-center text-bg-info text-white my-3 p-3">Register</h2>
-      <p id="error-message" class="text-danger fw-bold text-center"></p>
+      <p id="message-box" class="fw-bold text-center"></p>
       <form method="post" class="d-flex flex-column gap-3 col-8 col-md-4 col-lg-5 col-xl-5 col-xxl-5 m-auto" id="register-form">
         <div class="form-group">
           <label for="post-title">Username <span class="text-secondary">(must not contain punctuation symbols apart from underscore "_")</span></label>
@@ -48,6 +48,6 @@
     <footer>
       <p class="text-white-50 text-center text-bg-dark m-0 mt-4 p-5">SocialDock &#169; 2022</p>
     </footer>
-    <script src="/index.mjs" type="module"></script>
+    <script src="../src/js/register/index.mjs" type="module"></script>
   </body>
 </html>
