--- src/index.mjs
@@ -1,1 +1,0 @@
-import { registerUser } from "./js/register/register-user.mjs";
