+++ src/js/login/index.mjs
@@ -1,0 +1,1 @@
+import * as loginUser from "./loginUser.mjs";
