--- src/js/login/login.mjs
@@ -1,9 +1,0 @@
-import { API_BASE_URL } from "../constants/constants.mjs";
-
-async function loginUser(url, userData) {
-  try {
-    const response = await fetch(url, user);
-  } catch (error) {
-    console.log(error);
-  }
-}
