--- src/js/register/constants.mjs
@@ -1,1 +1,0 @@
-export const API_BASE_URL = "https://api.noroff.dev/api/v1";
