+++ src/js/register/index.mjs
@@ -1,0 +1,1 @@
+import * as registerUser from "./registerUser.mjs";
